Reinstate caps lock
Make the caps lock key send caps lock, relying on the OS to remap it to
ctrl.  Move the dual-function ctrl/esc to the bottom row.

This is to avoid typos, e.g. intending ctrl+tab but getting esc,tab.
This can be destructive (e.g. accidentally dismissing a dialog).